adding cleaned up UI finally
