+++ sdkproject/Assets/Mapbox/Unity/DataContainers/StreetsV7TileStats.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 5357c095b6c7e4315ae5edbc5540c7e5
+timeCreated: 1530546648
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
