+++ sdkproject/Assets/Mapbox/Unity/Editor/FeatureSubLayerTreeView.cs
@@ -11,18 +11,18 @@
 		public SerializedProperty Layers;
 		private float kToggleWidth = 18f;
 		public static int uniqueId = 3000;
-		const float kRowHeights = 20f;
+		const float kRowHeights = 15f;
 		const float nameOffset = 15f;
 		private GUIStyle columnStyle = new GUIStyle() 
 		{
-			alignment = TextAnchor.MiddleLeft, 
+			alignment = TextAnchor.MiddleCenter, 
 			normal = new GUIStyleState(){textColor = Color.white} 
 		};
 
 		public FeatureSubLayerTreeView(TreeViewState state, MultiColumnHeader multicolumnHeader, TreeModel<FeatureTreeElement> model) : base(state, multicolumnHeader, model)
 		{
 			// Custom setup
-			rowHeight = kRowHeights;
+			//rowHeight = kRowHeights;
 			showAlternatingRowBackgrounds = true;
 			showBorder = true;
 			customFoldoutYOffset = (kRowHeights - EditorGUIUtility.singleLineHeight) * 0.5f; // center foldout in the row since we also center content. See RowGUI
@@ -50,7 +50,6 @@
 
 			// Utility method that initializes the TreeViewItem.children and .parent for all items.
 			SetupParentsAndChildrenFromDepths(root, items);
-
 			// Return root of the tree
 			return root;
 		}
@@ -116,11 +115,11 @@
 			var rowItem = (TreeViewItem <FeatureTreeElement>)args.item;
 			for (int i = 0; i < args.GetNumVisibleColumns(); ++i)
 			{
-				CellGUI(args.GetCellRect(i), rowItem, (MyColumns)args.GetColumn(i), ref args);
+				CellGUI(args.GetCellRect(i), rowItem, (FeatureSubLayerColumns)args.GetColumn(i), ref args);
 			}
 		}
 
-		void CellGUI(Rect cellRect, TreeViewItem<FeatureTreeElement> item, MyColumns column, ref RowGUIArgs args)
+		void CellGUI(Rect cellRect, TreeViewItem<FeatureTreeElement> item, FeatureSubLayerColumns column, ref RowGUIArgs args)
 		{
 			// Center cell rect vertically (makes it easier to place controls, icons etc in the cells)
 			if (Layers == null || Layers.arraySize == 0)
@@ -135,7 +134,7 @@
 
 			var layer = Layers.GetArrayElementAtIndex(args.item.id - uniqueId);
 			CenterRectUsingSingleLineHeight(ref cellRect);
-			if (column == MyColumns.Name)
+			if (column == FeatureSubLayerColumns.Name)
 			{
 				Rect toggleRect = cellRect;
 				toggleRect.x += GetContentIndent(item);
@@ -155,7 +154,7 @@
 				//This draws the name property
 				base.RowGUI(args);
 			}
-			if (column == MyColumns.Type)
+			if (column == FeatureSubLayerColumns.Type)
 			{
 				cellRect.xMin += 15f; // Adding some gap between the checkbox and the name
 
@@ -166,7 +165,7 @@
 		}
 
 		// All columns
-		enum MyColumns
+		enum FeatureSubLayerColumns
 		{
 			Name,
 			Type
@@ -183,7 +182,8 @@
 					contextMenuText = "Name",
 					headerTextAlignment = TextAlignment.Center,
 					autoResize = true,
-					canSort = false
+					canSort = false,
+					allowToggleVisibility = false,
 				},
 
 				//Type column
@@ -193,87 +193,13 @@
 					contextMenuText = "Type",
 					headerTextAlignment = TextAlignment.Center,
 					autoResize = true,
-					canSort = false
+					canSort = false,
+					allowToggleVisibility = false
 				}
 			};
 
 			var state = new MultiColumnHeaderState(columns);
 			return state;
-		}
-	}
-}
-namespace Mapbox.Editor
-{
-	using System.Collections;
-	using System.Collections.Generic;
-	using UnityEngine;
-	using UnityEditor.IMGUI.Controls;
-	using UnityEditor;
-	using Mapbox.Unity.Map;
-
-	public class PointsOfInterestSubLayerTreeView : TreeView
-	{
-		public SerializedProperty Layers;
-		private float kToggleWidth = 18f;
-		private const int uniqueId = 0;//100000;
-
-		public PointsOfInterestSubLayerTreeView(TreeViewState state)
-			: base(state)
-		{
-			showAlternatingRowBackgrounds = true;
-			showBorder = true;
-			Reload();
-		}
-
-		protected override TreeViewItem BuildRoot()
-		{
-			// The root item is required to have a depth of -1, and the rest of the items increment from that.
-			var root = new TreeViewItem { id = -1, depth = -1, displayName = "Root" };
-
-			var items = new List<TreeViewItem>();
-			var index = 0;
-
-			if (Layers != null)
-			{
-				for (int i = 0; i < Layers.arraySize; i++)
-				{
-					var name = Layers.GetArrayElementAtIndex(i).FindPropertyRelative("coreOptions.sublayerName").stringValue;
-					items.Add(new TreeViewItem { id = index + uniqueId, depth = 1, displayName = name });
-					index++;
-				}
-			}
-
-			// Utility method that initializes the TreeViewItem.children and .parent for all items.
-			SetupParentsAndChildrenFromDepths(root, items);
-
-			// Return root of the tree
-			return root;
-		}
-
-		protected override bool CanRename(TreeViewItem item)
-		{
-			return true;
-		}
-
-		protected override void RenameEnded(RenameEndedArgs args)
-		{
-			if (Layers == null)
-			{
-				return;
-			}
-
-			var layer = Layers.GetArrayElementAtIndex(args.itemID - uniqueId);
-			layer.FindPropertyRelative("coreOptions.sublayerName").stringValue = string.IsNullOrEmpty(args.newName.Trim()) ? args.originalName : args.newName;
-		}
-
-		protected override void RowGUI(RowGUIArgs args)
-		{
-			Rect toggleRect = args.rowRect;
-			toggleRect.width = kToggleWidth;
-			var item = Layers.GetArrayElementAtIndex(args.item.id - uniqueId);
-			item.FindPropertyRelative("coreOptions.isActive").boolValue = EditorGUI.Toggle(toggleRect, item.FindPropertyRelative("coreOptions.isActive").boolValue);
-			args.item.displayName = item.FindPropertyRelative("coreOptions.sublayerName").stringValue;
-			base.RowGUI(args);
 		}
 	}
 }
