+++ sdkproject/Assets/Mapbox/Unity/Editor/PointsOfInterestSubLayerTreeView.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 430429020c0a64703b89ac33585da3e7
+timeCreated: 1530564120
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
